requirements.txt
asgiref==3.3.1
cffi==1.14.4
cryptography==3.4
Django==3.1.6
django-rest-knox==4.1.0
djangorestframework==3.12.2
psycopg2-binary==2.8.6
pycparser==2.20
pytz==2021.1
semantic-version==2.8.5
setuptools-rust==0.11.6
sqlparse==0.4.1
toml==0.10.2
